+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## What's new in Version 1.11.1 of RaspiBlitz?
 
 - Update: Specter Desktop 2.0.4 with reactivated UPDATE option [details](https://github.com/cryptoadvance/specter-desktop/releases/tag/v2.0.4)
+- Remove: Tallycoin-Connect [see service shutdown](https://x.com/djbooth007/status/1784409117563720082)
 
 ## What's new in Version 1.11.0 of RaspiBlitz?